Aufräumen und Inline
Unnötige dateien gelöscht und der Parser parsed jetzt
closes #20
