+++ PoE PartyTool Tests/PoE Party-Invite.csproj
@@ -55,7 +55,6 @@
     <Compile Include="PoE_PartyTool\Model\RequestSource.cs" />
     <Compile Include="PoE_PartyTool\Model\RequestSourceBuilder.cs" />
     <Compile Include="PoE_PartyTool\Utilities\PoEProcessWatcher.cs" />
-    <Compile Include="PoE_PartyTool\Utilities\RegexPatterns.cs" />
     <Compile Include="PoE_PartyTool\Utilities\ReverseLineReader.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
